+++ packages/icons/__tests__/shared.js
@@ -3,57 +3,42 @@
 import "react-native";
 import React from "react";
 import renderer from "react-test-renderer";
-import { colours } from "@times-components/styleguide";
-import { IconDiamond, IconTwitter, IconVideo } from "../src/icons";
+import { render, shallow } from "enzyme";
+import * as Icons from "../src/icons";
+
+function iconRenderTest(name, Icon) {
+  return () => {
+    const tree = renderer
+      .create(<Icon title={name} width={50} height={50} />)
+      .toJSON();
+
+    expect(tree).toMatchSnapshot();
+  };
+}
+
+function iconColourTest(Icon) {
+  return () => {
+    const stroke = "#coffee";
+    const fill = "#facade";
+
+    const tree = render(<Icon strokeColour={stroke} fillColour={fill} />);
+
+    expect(tree.find(`[stroke="${stroke}"]`).length > 0).toEqual(true);
+    expect(tree.find(`[fill="${fill}"]`).length > 0).toEqual(true);
+  };
+}
+
+function iconDimensionTest(Icon) {
+  return () => {
+    const tree = shallow(<Icon height={50} />);
+    expect(tree.find("Svg").props().width > 0).toEqual(true);
+  };
+}
 
 module.exports = () => {
-  it("Diamond renders correctly", () => {
-    const tree = renderer
-      .create(
-        <IconDiamond
-          width={50}
-          height={50}
-          fillColour={colours.functional.tertiary}
-        />
-      )
-      .toJSON();
-
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("Twitter renders correctly", () => {
-    const tree = renderer
-      .create(<IconTwitter width={50} height={50} />)
-      .toJSON();
-
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("Twitter renders correctly with different fillColour", () => {
-    const tree = renderer
-      .create(
-        <IconTwitter
-          width={50}
-          height={50}
-          fillColour={colours.functional.tertiary}
-        />
-      )
-      .toJSON();
-
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("Video renders correctly", () => {
-    const tree = renderer.create(<IconVideo width={50} height={50} />).toJSON();
-
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("Video renders correctly with different fillColour", () => {
-    const tree = renderer
-      .create(<IconVideo width={50} height={50} fillColour="#4D4D4D" />)
-      .toJSON();
-
-    expect(tree).toMatchSnapshot();
+  Object.entries(Icons).forEach(([name, Icon]) => {
+    it(`${name} renders correctly`, iconRenderTest(name, Icon));
+    it(`${name} renders with different colours`, iconColourTest(Icon));
+    it(`${name} sets a width when height is set`, iconDimensionTest(Icon));
   });
 };